feat: do not override Node.js require (fixes #8)
BREAKING CHANGE:

It occured to me that client-side applications are designed with a single-entry script. There is no need to override “require”. It is enough to evaluate the entry script.

If I am wrong, it is easy to revert this change, i.e. add ability to require individual paths in addition to evaluating the entry script.